Update keyboardShortcut.js
hey bro! this is not my documentation i just forked to test this and found some error as you are saying about the incorrectly calls Spicetify.Player.prev, a non-existent function, instead of Spicetify.Player.back.

Solves #3122, which was closed without proper testing of the extension. i have fixed this in my forked one i can not do this in the orignal documentation...